pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diabimmune-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "hydra-core>=1.3.0",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "trackio>=0.3.0",
]

[tool.setuptools]
packages = ["modules", "utils"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
